Move scan and walk into own files
